change userController.ts, userModel.ts, usersData.ts
