+++ src/userController.ts
@@ -1,7 +1,6 @@
 import { IncomingMessage, ServerResponse } from 'http';
-import * as User from './userModel'
-import { validate } from 'uuid';
-import { IUser, IUserWithoutId, users } from './usersData';
+import * as User from './userModel';
+import { IUserWithoutId } from './usersData';
 import { getPostData } from './utils';
 
 async function getAllUsers(req: IncomingMessage, res: ServerResponse): Promise<void> {
@@ -17,17 +16,9 @@ async function getAllUsers(req: IncomingMessage, res: ServerResponse): Promise<v
 
 async function getUserById(req: IncomingMessage, res: ServerResponse, id: string): Promise<void> {
     try {
-        const user = await User.findById(id);
-        if (id && !validate(id)) {
-            res.writeHead(400, {'Content-Type': 'application/json'});
-            res.end(JSON.stringify({ message: 'Invalid id' }));
-        } else if (!user) {
-            res.writeHead(404, {'Content-Type': 'application/json'});
-            res.end(JSON.stringify({ message: 'User Not Found' }));
-        } else {
-            res.writeHead(200, {'Content-Type': 'application/json'});
-            res.end(JSON.stringify(user));
-        };
+        const user = await User.findById(res, id);
+        res.writeHead(200, {'Content-Type': 'application/json'});
+        res.end(JSON.stringify(user));
     }catch(error) {
         res.writeHead(500, {'Content-Type': 'application/json'});
         res.end(JSON.stringify({message: 'Server down'}));
@@ -47,32 +38,24 @@ async function createUser(req: IncomingMessage, res: ServerResponse): Promise<vo
         res.writeHead(201, {'Content-Type': 'application/json'});
         res.end(JSON.stringify(newUser));
     } catch(error) {
-        res.writeHead(400, {'Content-Type': 'application/json'});
-        res.end(JSON.stringify({ message: 'Missing required fields' }));
+        res.writeHead(500, {'Content-Type': 'application/json'});
+        res.end(JSON.stringify({message: 'Server down'}));
     };
 };
 
 async function updateUser(req: IncomingMessage, res: ServerResponse, id: string): Promise<void> {
     try {
-        const user = await User.findById(id);
-        if (id && !validate(id)) {
-            res.writeHead(400, {'Content-Type': 'application/json'});
-            res.end(JSON.stringify({ message: 'Invalid id' }));
-        } else if (!user) {
-            res.writeHead(404, {'Content-Type': 'application/json'});
-            res.end(JSON.stringify({ message: 'User Not Found' }));
-        } else {
-            const body = await getPostData(req, res);
-            const { username, age, hobbies }: IUserWithoutId = body;
-            const userData: IUserWithoutId = {
-                username: username,
-                age: age,
-                hobbies: hobbies
-            };
-            const updUser = await User.update(id, userData);
-            res.writeHead(200, {'Content-Type': 'application/json'});
-            res.end(JSON.stringify(updUser));
+        await User.findById(res, id);
+        const body = await getPostData(req, res);
+        const { username, age, hobbies }: IUserWithoutId = body;
+        const userData: IUserWithoutId = {
+            username: username,
+            age: age,
+            hobbies: hobbies
         };
+        const updUser = await User.update(id, userData);
+        res.writeHead(200, {'Content-Type': 'application/json'});
+        res.end(JSON.stringify(updUser));
     }catch(error) {
         res.writeHead(500, {'Content-Type': 'application/json'});
         res.end(JSON.stringify({message: 'Server down'}));
@@ -81,18 +64,10 @@ async function updateUser(req: IncomingMessage, res: ServerResponse, id: string)
 
 async function deleteUser(req: IncomingMessage, res: ServerResponse, id: string): Promise<void> {
     try {
-        const user = await User.findById(id);
-        if (id && !validate(id)) {
-            res.writeHead(400, {'Content-Type': 'application/json'});
-            res.end(JSON.stringify({ message: 'Invalid id' }));
-        } else if (!user) {
-            res.writeHead(404, {'Content-Type': 'application/json'});
-            res.end(JSON.stringify({ message: 'User Not Found' }));
-        } else {
-            await User.remove(id, users);
-            res.writeHead(204, {'Content-Type': 'application/json'});
-            res.end(JSON.stringify({ message: `User ${id} removed` }));
-        };
+        await User.findById(res, id);
+        await User.remove(id);
+        res.writeHead(204, {'Content-Type': 'application/json'});
+        res.end(JSON.stringify({ message: `User ${id} removed` }));
     }catch(error) {
         res.writeHead(500, {'Content-Type': 'application/json'});
         res.end(JSON.stringify({message: 'Server down'}));
