+++ src/usersData.ts
@@ -1,7 +1,3 @@
-import { v4 as uuidv4 } from 'uuid';
-
-export let users: IUser[] = [];
-
 interface IUser {
     id: string;
     username: string;
@@ -14,15 +10,6 @@ interface IUserWithoutId {
     age: number;
     hobbies: string[] | never[];
 }
-
-const userOne: IUser = {
-    id: uuidv4(),
-    username: 'Vasya',
-    age: 30,
-    hobbies: []
-}
-
-users.push(userOne);
 
 export {
     IUser,
